no long long

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,7 +7,7 @@
 int main(void)
 {
 	int i, l;
-	unsigned int a, b, c;
+	long a, b, c;
 
 	a = 1;
 	b = 2;
